Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

Removed Android.mk, CleanSpec.mk and egl.cfg

Also did some changes in the kernel config, and now the OS kernel works with recovery as well, using PURE_ANDROID.

Btw, I've noticed that with the backported USB drivers, the USB file transfer speed is lower than before. It used to be around 3 MB/s, but now it's just around 1.3 MB/s. Is it the same on other devices as well, with the backported drivers?

Patch-set: 3