Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

CUSTOM_LOCALES seems to affect the build for ALL devices, which are not good. So switch to PRODUCT_LOCALES instead. On a fresh install this will set the default locale to a non-valid one, until the user manually changes the language, but I don't know any other way to do this atm.

Patch-set: 2