Update patch set 4
Patch Set 4:

Default framebuffer BPP in the msm7227 kernels we've been based on, are using 32, while the Wildfire should use a BPP of 16. I just discovered that error while comparing some stuff. Although it seemed to work ok with 32, I figured it was best to replace it with the value that the Wildfire is supposed to use.

Btw could we get some merge love around here? Lot's of stuff waiting in the queue. :)

Patch-set: 4